removed-testspec-code
Great work! Keep in mind, your "solution" file runs first, this is where you define all the functions needed for your test specs.  You don't have to invoke, call, or pass arguments to the functions in your properties.js file, this is what the test spec does for you! The specs will invoke and use the code you defined in properties.js and use inside of each spec.  In your specs, you will see code that looks like:

```
var someArray = [];
expect(setPropertiesOnArrayObj(someArray)).toEqual(someArray);
```

Inside of the `expect` the function you defined is invoked and passed an argument, then `toEqual` checks the value of `someArray` setPropertiesOnArrayObj is invoked!

Great work! When you get a chance, please use the forked version of this repository, it makes grading and adding comments a lot easier for instructors.  The process of doing this is:

1. Fork Fullstack Test First from the Fullstack's repository.
2. Clone the forked version that is now on your account to your local machine and use that version.
3. The repo will say it is private but since you forked it from Fullstack's account, we can access it and make comments on it.

Great work!